Add function Printing Fiche
